Move bank benchmark outside src
This will make it available to third party benchmarking tools.
